+++ src/context.rs
@@ -340,12 +340,12 @@ fn update<S: ScheduleLabel>(
 
         let gamepad = context.get::<GamepadDevice>().copied().unwrap_or_default();
         let context_active = instance.is_active(&context.as_readonly());
-        let Some(context_actions) = instance.actions_mut(&mut context) else {
+        let Some(mut context_actions) = instance.actions_mut(&mut context) else {
             continue;
         };
 
-        context_actions.sort_by_cached_key(|&action| {
-            let Ok((.., action_bindings, _, _, _)) = actions.get(action) else {
+        let mods_count = |action: &Entity| {
+            let Ok((.., action_bindings, _, _, _)) = actions.get(*action) else {
                 // TODO: use `warn_once` when `bevy_log` becomes `no_std` compatible.
                 warn!(
                     "`{action}` from `{}` missing action components",
@@ -360,13 +360,17 @@ fn update<S: ScheduleLabel>(
                 .max()
                 .unwrap_or(0);
             Reverse(value)
-        });
+        };
+
+        if !context_actions.is_sorted_by_key(mods_count) {
+            context_actions.sort_by_cached_key(mods_count);
+        }
 
         trace!("updating `{}` on `{}`", instance.name, instance.entity);
 
         reader.set_gamepad(gamepad);
 
-        let mut actions_iter = actions.iter_many_mut(context_actions);
+        let mut actions_iter = actions.iter_many_mut(&*context_actions);
         while let Some((
             action,
             action_name,